+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1.3.2
 ---------------------------------------------------------------
 * **cloak\result\Coverage** to **cloak\value\Coverage**
 * add **cloak\value\LineSet** class
+* add **cloak\value\LineRange** class
 
 Version 1.3.1
 ---------------------------------------------------------------